refactor(rf_model.py): rename model file
